Added a pointer to the message structure.
This example shows how to deal with a pointer in the message structure.
The size of the message just includes the size of the pointer. We need
to copy the original data into a new buffer allocated at the location
that pointer points to, then we need to copy the original data there.
This works because the stats thread can see the same new memory
location.